extract main() to main.rs
